% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{articles}
\alias{articles}
\title{articles}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 23 rows and 2 columns.
}
\usage{
articles
}
\description{
Data.frame with common articles
}
\seealso{
clean_strings
}
\keyword{datasets}
