% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_clean_settings.R
\name{build_clean_settings}
\alias{build_clean_settings}
\title{Building settings for string cleaning}
\usage{
build_clean_settings(
  sp_char_words = fedmatch::sp_char_words,
  common_words = NULL,
  remove_char = NULL,
  remove_words = FALSE,
  stem = FALSE,
  replace_null = NULL
)
}
\arguments{
\item{sp_char_words}{character vector. Data.frame where first column is special characters and second column is full words. The default is}

\item{common_words}{data.frame. Data.frame where first column is abbreviations and second column is full words.}

\item{remove_char}{character vector. string of specific characters (for example, "letters") to be removed}

\item{remove_words}{logical. If TRUE, removes all abbreviations and replacement words in common_words}

\item{stem}{logical. If TRUE, words are stemmed}

\item{replace_null}{character vector. If not NULL, the value with which to replace empty or blank strings.}
}
\value{
list with settings to pass to \code{clean_strings}
}
\description{
\code{build_score_settings} is a convenient way to make the proper list for the
\code{clean_settings} argument of \code{merge_plus}.
}
