% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_corpus.R
\name{build_corpus}
\alias{build_corpus}
\title{Calculate word corpus for weighted jaccard matching}
\usage{
build_corpus(namelist1, namelist2)
}
\arguments{
\item{namelist1}{character vector of names from dataset 1}

\item{namelist2}{character vector of names from dataset 2}
}
\value{
a data.table with columns for frequency, inverse frequency, and log inverse frequency for each word in the two strings.
}
\description{
Calculate word corpus for weighted jaccard matching
}
