% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fuzzy_settings.R
\name{build_fuzzy_settings}
\alias{build_fuzzy_settings}
\title{Build settings for fuzzy matching}
\usage{
build_fuzzy_settings(
  method = "jw",
  p = 0.1,
  maxDist = 0.05,
  matchNA = FALSE,
  nthread = getOption("sd_num_thread")
)
}
\arguments{
\item{method}{character vector of length 1. Either one of the methods listed in stringdist::amatch, or our custom method 'wgt_jaccard.' See the vignettes for more details.}

\item{p}{numeric vector of length 1. See stringdist::amatch()}

\item{maxDist}{numeric vector of length 1. See stringdist::amatch()}

\item{matchNA}{whether or not to match on NAs, see stringdist::amatch()}

\item{nthread}{number of threads to use in the underlying C code.}
}
\description{
\code{build_fuzzy_settings} is a convenient way to build the list for the fuzzy settings argument in merge_plus
}
