% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_score_settings.R
\name{build_score_settings}
\alias{build_score_settings}
\title{Build settings for scoring}
\usage{
build_score_settings(
  score_var_x = NULL,
  score_var_y = NULL,
  score_var_both = NULL,
  wgts = NULL,
  score_type
)
}
\arguments{
\item{score_var_x}{character vector. the variables from the 'x' dataset to score on}

\item{score_var_y}{character vector. the variables from the 'y' dataset to score on}

\item{score_var_both}{the variables from both datasets (shared names) to score on, before any prefixes are applied.}

\item{wgts}{numeric vector. The weights for the linear sum of scores}

\item{score_type}{character vector.}
}
\description{
\code{build_score_settings} is a convenient way to make the proper list for the
\code{score_settings} argument of \code{merge_plus} Each vector in build_score_settings
should be the same length, and each position (first, second, third, etc.)
corresponds to one variable to score on.
}
