% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{corp_data2}
\alias{corp_data2}
\title{corp_data2}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 6 columns.
}
\usage{
corp_data2
}
\description{
Some made up data on the top 10 US companies in the Fortune 500. Mock-matched to corp_data1 in examples/match_template.R
}
\keyword{datasets}
