% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{corporate_words}
\alias{corporate_words}
\title{corporate_words}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 54 rows and 2 columns.
}
\usage{
corporate_words
}
\description{
Data.frame with common corporate abbreviations in column 1 and corresponding long names in column 2. Useful for cleaning company names for matching.
}
\seealso{
clean_strings
}
\keyword{datasets}
