% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data1}
\alias{data1}
\title{data1}
\format{An object of class \code{data.frame} with 10 rows and 5 columns.}
\usage{
data1
}
\description{
Some made up data on the top 10 US companies in the Fortune 500. Mock-matched to data2 in examples/match_template.R
}
\keyword{datasets}
