% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_eliminate.R
\name{duplicate_eliminate}
\alias{duplicate_eliminate}
\title{duplicate_eliminate}
\usage{
duplicate_eliminate(data, id_vars, dupout = FALSE, tag = NULL)
}
\arguments{
\item{data}{data.frame}

\item{id_vars}{character vector.}

\item{dupout}{logical. If FALSE, returns dataset without duplicates. If TRUE, returns duplicates.}
}
\value{
see dupout
}
\description{
Identifies and eliminates all duplicates including the first instance, if they exist
}
\details{
h/t Simon O'Hanlon and his Stack Overflow post
\url{http://stackoverflow.com/questions/17352657/using-r-how-can-i-flag-sequential-duplicate-values-in-a-single-column-of-a-data}
}
\examples{
data = data.frame('a'=c(1,1,1,1,2,2,2,2,3), 'b'=c(1,1,2,4,1,2,2,3,1))
duplicate_eliminate(data,c('a','b'))
duplicate_eliminate(data,c('a'))


}
