% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{fund_words}
\alias{fund_words}
\title{fund_words}
\format{
An object of class \code{data.frame} with 63 rows and 2 columns.
}
\usage{
fund_words
}
\description{
Data.frame with abbreviations common in the names of financial (i.e. mutual) funds in column 1 and corresponding long names in column 2.  Useful for cleaning fund names for matching.
}
\seealso{
clean_strings
}
\keyword{datasets}
