% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sp_char_words}
\alias{sp_char_words}
\title{sp_char_words}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
sp_char_words
}
\description{
Common special characters and their replacements for string cleaning
}
\keyword{datasets}
