% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{State_FIPS}
\alias{State_FIPS}
\title{State_FIPS}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 50 rows and 2 columns.
}
\usage{
State_FIPS
}
\description{
State FIPS codes
}
\keyword{datasets}
