% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsor.R
\name{winsor}
\alias{winsor}
\title{winsor}
\usage{
winsor(x, fraction = 0.05)
}
\arguments{
\item{x}{numeric or vector. variable to be winsored}

\item{fraction}{numeric. winsor cutoff}
}
\value{
x winsored at fraction
}
\description{
winsor
}
\examples{


}
