% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_frequency.R
\name{word_frequency}
\alias{word_frequency}
\title{word_frequency}
\usage{
word_frequency(string)
}
\arguments{
\item{string}{character vector}
}
\value{
data frame with word frequency
}
\description{
A function to count occurences in a set of strings; function does minimal cleaning (remove punctuation and extra spaces);
}
\examples{
# with simple vector
names = c("company name 1", "company name 2", "company 3")
word_frequency(names)

# with a dataframe
col1 = c(1,2,3)
df1 = cbind(col1, names)
word_frequency(df1[,2])

# more complicated example
names = c("company name 1", "company &    3 inc.", "co nm 1,, lp")
df2 = cbind(col1, names)
word_frequency(df2[,2])


}
