% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{World_Bank_Codes}
\alias{World_Bank_Codes}
\title{World_Bank_Codes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 213 rows and 2 columns.
}
\usage{
World_Bank_Codes
}
\description{
World Bank 3-Character Country Codes for 213 countries
}
\keyword{datasets}
