% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean_strings}
\alias{clean_strings}
\title{clean_strings}
\usage{
clean_strings(string, sp_char_words = NULL, common_words = NULL,
  remove_char = NULL, remove_words = FALSE, stem = FALSE,
  replace_null = NULL)
}
\arguments{
\item{string}{character or character vector of strings}

\item{sp_char_words}{character vector. Data.frame where first column is special characaters and second column is full words.}

\item{common_words}{data.frame. Data.frame where first column is abbreviations and second column is full words.}

\item{remove_char}{character vector. string of specific characters (for example, "letters") to be removed}

\item{remove_words}{logical. If TRUE, removes all abbreviations and replacement words in common_words}

\item{stem}{logical. If TRUE, words are stemmed}
}
\value{
cleaned strings
}
\description{
default string cleaning process for "name_match"
}
\examples{
# basic cleaning example
sample_str = c("Holding Co. A @ B St, 3 ))", "Company B & C 4, Inc.", "make $, inc.")
sample_clean1 = clean_strings(sample_str)

# defining common words with a package database
sample_clean2 = clean_strings(sample_str, common_words=corporate_words)

# dropping common words in a database
sample_clean3 = clean_strings(sample_str, common_words=corporate_words, remove_words=TRUE)

# sunco example
sample_clean4 = clean_strings("co cosuncosunco co co", common_words = cbind(c("co"), c("company")))

# changing special characters to words(Note that @ and & are dropped with punctuation)
drop_char = cbind(c("\\\\$", "\\\\\%"), c("dollar", "percent"))
sample_clean5 = clean_strings(sample_str, sp_char_words = drop_char)


}
