% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_frequency.R
\name{word_frequency}
\alias{word_frequency}
\title{Compute frequency of words in a corpus}
\usage{
word_frequency(string)
}
\arguments{
\item{string}{character vector}
}
\value{
data.table with word frequency
}
\description{
\code{word_frequency} counts the frequency of words in a set of strings.
Also does minimal cleaning (removes punctuation and extra spaces). Useful for
determining what words are common and may need to be replaced or removed with
\code{clean_strings}.
}
